lpe/* Makefile.in: - moved to trunk/src
configure:
- renamed `--enable-experimentals' to `--enable-experimental' (on popular demand)



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@824 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/studio/studio_overview.cpp
@@ -366,7 +366,7 @@ void StudioOverview::AddSpecifications(wxCommandEvent &event) {
   file_helper.ClearExt();
 
   /* A new basename */
-  std::string name = std::string(file_helper.GetFullName().Append(wxString::Format(wxT("-%u"), specifications->GetChildrenCount(selected, false))).fn_str());
+  std::string name = std::string(file_helper.GetFullName().Append(wxString::Format(wxT("-%u"), (unsigned int) specifications->GetChildrenCount(selected, false))).fn_str());
 
   /* Connect input object to compatible position (TODO generalise to multiple inputs and outputs) */
   std::vector < SpecificationInputType >& input_objects = new_specification->GetModifiableInputObjects();
